deps: Add new react-select lib and deprecate the old one
