+++ components/EditWebhooks.js
@@ -17,7 +17,7 @@ import StyledHr from './StyledHr';
 import MessageBox from './MessageBox';
 import { Flex, Box } from '@rebass/grid';
 import StyledButton from './StyledButton';
-import StyledSelect from './StyledSelect';
+import StyledSelect from './DeprecatedStyledSelect';
 import { Add } from 'styled-icons/material/Add';
 import StyledInputGroup from './StyledInputGroup';
 import ExternalLinkNewTab from './ExternalLinkNewTab';
