+++ lib/constants/intervals.js
@@ -1,0 +1,6 @@
+const INTERVALS = {
+  month: 'month',
+  year: 'year',
+};
+
+export default INTERVALS;
