+++ lib/constants/theme.js
@@ -28,24 +28,28 @@ export const defaultColors = {
     100: '#E6FAEF',
   },
   blue: {
-    800: '#0041A3',
-    700: '#145ECC',
-    500: '#3385FF',
-    400: '#66A3FF',
-    300: '#99C9FF',
-    200: '#B8DEFF',
-    100: '#EBF4FF',
-    50: '#F0F8FF',
+    900: '#0C2D66',
+    800: '#1041A3',
+    700: '#1153D6',
+    600: '#1869F5',
+    500: '#297EFF',
+    400: '#5CA3FF',
+    300: '#8FC7FF',
+    200: '#C2E2FF',
+    100: '#E6F3FF',
+    50: '#F5FAFF',
   },
   primary: {
-    800: '#0041A3',
-    700: '#145ECC',
-    500: '#3385FF',
-    400: '#66A3FF',
-    300: '#99C9FF',
-    200: '#B8DEFF',
-    100: '#EBF4FF',
-    50: '#F0F8FF',
+    900: '#0C2D66',
+    800: '#1041A3',
+    700: '#1153D6',
+    600: '#1869F5',
+    500: '#297EFF',
+    400: '#5CA3FF',
+    300: '#8FC7FF',
+    200: '#C2E2FF',
+    100: '#E6F3FF',
+    50: '#F5FAFF',
   },
   red: {
     700: '#CC1836',
