+++ package.json
@@ -83,6 +83,7 @@
     "react-mde": "6.0.0",
     "react-quill": "1.3.3",
     "react-scrollchor": "6.0.0",
+    "react-select": "3.0.4",
     "react-stickynode": "2.1.1",
     "react-stripe-elements": "4.0.0",
     "react-tag-input": "6.4.0",
