+++ pages/discover.js
@@ -15,7 +15,7 @@ import { H1, P } from '../components/Text';
 import LoadingGrid from '../components/LoadingGrid';
 import Pagination from '../components/Pagination';
 import MessageBox from '../components/MessageBox';
-import StyledSelect from '../components/StyledSelect';
+import StyledSelect from '../components/DeprecatedStyledSelect';
 import { Link } from '../server/pages';
 import SearchForm from '../components/SearchForm';
